Add abstract DAO calss
